test(core): move WebGL mark tests behind renderer boundary
Keep semantic mark tests focused on shared encoding and configuration behavior.

Relocate renderer callbacks, uniforms, and shader snapshots beside the WebGL implementation.
